Change round_trip_time type to f32 according C++ float type

+++ rust/src/router/consumer.rs
@@ -239,7 +239,7 @@ pub struct ConsumerStat {
     pub packet_count: usize,
     pub byte_count: usize,
     pub bitrate: u32,
-    pub round_trip_time: Option<f64>,
+    pub round_trip_time: Option<f32>,
 }
 
 /// RTC statistics of the consumer, may or may not include producer statistics.
